requirements-runtime.txt
# Core Runtime
requests==2.32.3
pydantic>=2.5,<3
pypdf==5.1.0
pdfplumber>=0.10.0
openpyxl==3.1.5
pyyaml==6.0.2
jinja2==3.1.5
sentence-transformers==3.4.1
chromadb>=0.5.0
beautifulsoup4==4.12.3
defusedxml>=0.7.1
rank-bm25>=0.2.2
tqdm>=4.66.0
numpy>=1.24.0
networkx>=3.1
json-repair>=0.30.0

# MCP Server
mcp>=0.1.0,<2.0
